Backend/requirements.txt
flask
flask-cors
phi
python-dotenv
whisper
moviepy
